Separated core implementation and exampled code into two different projects.
