+++ core/src/main/java/me/ramswaroop/jbot/slackbot/core/BotWebSocketHandler.java
