+++ core/src/main/java/me/ramswaroop/jbot/slackbot/core/SlackService.java
