+++ core/src/main/java/me/ramswaroop/jbot/slackbot/core/models/DndStatus.java
