+++ core/src/main/java/me/ramswaroop/jbot/slackbot/core/models/RTM.java
