+++ example/src/main/java/me/ramswaroop/jbot/JBotApplication.java
