+++ example/src/main/java/me/ramswaroop/jbot/slackbot/SlackBot.java
