+++ example/src/main/java/me/ramswaroop/jbot/slackbot/SlackSlashCommand.java
